requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
Pillow
psycopg2-binary
PyJWT
python-dotenv
pytz
sqlparse
drf-yasg